requirements.prod.txt
PyJWT==0.2.1
django-cors-headers==0.12
djangorestframework==2.3.13
djangorestframework-jwt==0.1.5
